Detalles de nombres arreglados
Tambien se añadio botones de siguiente y anterior en cada leccion
