+++ uniplus/app/components/LandingPageScroll.jsx
@@ -1,10 +1,10 @@
-
+import { Video } from "@/app/components/video";
 
 export default function LandingPageScroll(){
     return (
         <>
           <div className="bg-[#0d2e4e] w-4/5 rounded-2xl mx-auto h-80  mb-10 text-gray-200">
-            Scroll de js con info sin definir, videos etc.
+            <Video></Video>
             </div>
         </>
     );
