+++ uniplus/app/components/lessonRenderer.jsx
@@ -7,20 +7,20 @@ import Link from "next/link";
 import { cursosData } from "@/app/data/cursosData";
 import { useRouter } from "next/navigation";
 
-
 export default function LessonRenderer({
   contenido,
   idCurso,
   idModulo,
   idLeccion,
-  moduleKey
+  moduleKey,
 }) {
   const router = useRouter();
   const totalLecciones = cursosData[idCurso].modulos[idModulo].lecciones.length;
   const esUltimaLeccion = idLeccion === totalLecciones - 1;
+  const esPrimeraLeccion = idLeccion === 0;
 
-  async function handleStartLesson (e) {
-    e.preventDefault(); 
+  async function handleStartLesson(e) {
+    e.preventDefault();
     const moduleUrl = `/courses/${idCurso}/${idModulo}/${idLeccion + 1}`;
     try {
       const response = await fetch("/api/updateProgress", {
@@ -38,9 +38,30 @@ export default function LessonRenderer({
     } catch (error) {
       console.error("Error al registrar el progreso del módulo:", error);
     }
-  };
-  async function handleStartExam (e) {
-    e.preventDefault(); 
+  }
+
+  async function handleBackLesson(e) {
+    e.preventDefault();
+    const moduleUrl = `/courses/${idCurso}/${idModulo}/${idLeccion - 1}`;
+    try {
+      const response = await fetch("/api/updateProgress", {
+        method: "PUT",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({ moduleKey, moduleUrl }),
+      });
+
+      if (!response.ok) {
+        throw new Error("Error al registrar el progreso del módulo");
+      }
+      router.push(moduleUrl);
+    } catch (error) {
+      console.error("Error al registrar el progreso del módulo:", error);
+    }
+  }
+  async function handleStartExam(e) {
+    e.preventDefault();
     const moduleUrl = `/courses/${idCurso}/${idModulo}/${idLeccion}/${idModulo}`;
     try {
       const response = await fetch("/api/updateProgress", {
@@ -58,7 +79,7 @@ export default function LessonRenderer({
     } catch (error) {
       console.error("Error al registrar el progreso del módulo:", error);
     }
-  };
+  }
   return (
     <div className="relative w-full flex justify-center items-center z-0">
       <CloudsBg />
@@ -147,15 +168,28 @@ export default function LessonRenderer({
           }
           return null;
         })}
+        <div className="flex gap-20">
+        {!esPrimeraLeccion && (
+          <a
+            href={`/courses/${idCurso}/${idModulo}/${idLeccion - 1}`}
+            onClick={handleBackLesson}
+            className="bg-[#4a7298] text-white text-2xl px-6 py-3 rounded-lg mt-6 font-bold hover:bg-[#3a5a78]"
+          >
+            Anterior
+          </a>
+        )}
         {!esUltimaLeccion && (
           <a
             href={`/courses/${idCurso}/${idModulo}/${idLeccion + 1}`}
             onClick={handleStartLesson}
             className="bg-[#4a7298] text-white text-2xl px-6 py-3 rounded-lg mt-6 font-bold hover:bg-[#3a5a78]"
           >
-            Ir a la siguiente Lección
+            Siguiente
           </a>
         )}
+
+        </div>
+
       </div>
     </div>
   );
