+++ uniplus/app/components/navbar.jsx
@@ -22,7 +22,7 @@ export default function NavBar(){
                 /> ``
               </div>
               <div className="text-2xl flex items-center gap-20 text-[#2f66a5]">
-                <Link href="/contacto" className={`flex items-center gap-2 cursor-pointer hover:text-[#85baf7] ${pathName==="/contacto"? "text-[#092c53] font-bold":""}`}>
+                <Link href="/perfil" className={`flex items-center gap-2 cursor-pointer hover:text-[#85baf7] ${pathName==="/contacto"? "text-[#092c53] font-bold":""}`}>
                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={2} stroke="currentColor" className="size-7 ">
                   <path strokeLinecap="round" strokeLinejoin="round" d="M15 9h3.75M15 12h3.75M15 15h3.75M4.5 19.5h15a2.25 2.25 0 0 0 2.25-2.25V6.75A2.25 2.25 0 0 0 19.5 4.5h-15a2.25 2.25 0 0 0-2.25 2.25v10.5A2.25 2.25 0 0 0 4.5 19.5Zm6-10.125a1.875 1.875 0 1 1-3.75 0 1.875 1.875 0 0 1 3.75 0Zm1.294 6.336a6.721 6.721 0 0 1-3.17.789 6.721 6.721 0 0 1-3.168-.789 3.376 3.376 0 0 1 6.338 0Z" />
                 </svg>
