+++ uniplus/app/data/questionsData.js
@@ -1,0 +1,83 @@
+export const questionsData = [
+    {
+      "question": "La capital de Australia es Sídney.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "El ser humano tiene 206 huesos.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La Gran Muralla China es visible desde la Luna.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "La fórmula química del agua es H2O.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La Tierra es el tercer planeta del sistema solar.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "Albert Einstein ganó el Premio Nobel por su teoría de la relatividad.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "Los humanos y los dinosaurios coexistieron.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "La capital de Canadá es Toronto.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "El océano más grande del mundo es el Océano Pacífico.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La Mona Lisa fue pintada por Leonardo da Vinci.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La Gran Pirámide de Giza está en Egipto.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "El oro es un metal no conductor de electricidad.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "La capital de Brasil es Río de Janeiro.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "Los murciélagos son mamíferos.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La lengua oficial de Suiza es solo el alemán.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "El monte Everest es la montaña más alta del mundo.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "El ser humano puede distinguir más de un millón de colores.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "El Sol es un planeta.",
+      "correct_answer": "False"
+    },
+    {
+      "question": "La Antártida es el continente más seco del mundo.",
+      "correct_answer": "True"
+    },
+    {
+      "question": "La Segunda Guerra Mundial terminó en 1945.",
+      "correct_answer": "True"
+    }
+  ];
+
